port: only unplug/stop if necessary
When reconfiguring an interface port, port_unplug() was called
unconditionally even when the requested changes did not require it.

Add an early return if none of the port-specific attributes are being
modified (number of RX/TX queues, queue sizes, or MAC address). This
allows changing interface flags and MTU without unplugging the port.

Signed-off-by: Robin Jarry <rjarry@redhat.com>