infra: add automatic rule and route setup for local VRF traffic
When a VRF is created, a corresponding gr-loopX interface is also created
in Linux.

To ensure that local traffic is properly routed through gr-loopX, the following
rules and routes must be added on linux side:

    ip rule add oif gr-loopX lookup 100X
    ip route add default dev gr-loopX table 100X
    ip rule add iif gr-loopX lookup 200X
    ip route add local default dev gr-loopX table 200X

This commit adds the logic to automatically install and remove these
rules and routes when a VRF is created or deleted.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>