frr: expose only the linux ifindex to frr
Some daemons (such as ISISd) rely on the ifindex to open a raw socket.
Yet, the registered ifindex in FRR is the Grout internal id, which
has no existence in Linux.
Add a mapping table between grout ifindex and the controlplane
tap ifindex.
This code may find a better place in grout directly, to extend
the usage to other control plane applications transparently.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>