infra: rename gr-loopX to gr-vrfX
The next commit will add a VRF interface on top of gr-loopX on the kernel
side. This is required to route packets correctly for daemons such as
bgpd that receive traffic on gr-loop1 and must send replies back through
the same interface.

By default, daemons like bgpd listen on 0.0.0.0, meaning their sockets
are not bound to any specific device.

To ensure that packets received on gr-loop1 are sent back through
gr-loop1, we need a dedicated routing context per interface.

There are two possible approaches:
– use a dedicated routing table selected by a policy rule (based on oif)
– or, use a dedicated VRF device with its own routing table.

The policy rule method cannot be used here, because rules are evaluated
before route lookup, and sockets not bound to an interface have no oif
value at that stage. Therefore, a separate vrf-X interface with a default
route via its corresponding gr-loopX is the only reliable option.

Thanks to tcp_l3mdev_accept, any new TCP connection created after
receiving a SYN on a gr-loopX device will automatically inherit the VRF
context of that device, ensuring correct reply routing.

As a VRF must be created on top of each gr-loopX, the interface name
cannot be shared between the VRF and the underlying TUN device.

At the same time, it is important that the interface name exposed to
zebra matches the one used in Grout.

To simplify this, the interface in Grout is renamed from gr-loopX to
vrf-X, while the TUN device keeps the original gr-loopX name.

Because gr-loop0 corresponds to the primary namespace, no VRF is needed
for it. It therefore remains named gr-loop0 both for the Grout and TUN
interfaces.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>
Reviewed-by: Christophe Fontaine <cfontain@redhat.com>
Acked-by: Robin Jarry <rjarry@redhat.com>