infra: create VRF and default route for each loopback interface
This commit introduces automatic VRF and route setup for local loopback
interfaces created by Grout.

When a new VRF is added in Grout, a corresponding Linux VRF interface is
now created and linked with its dedicated loopback device.

For each VRF vrf-X, a Linux VRF interface with the same name (vrf-X) is
created with routing table ID 1000 + X.

The associated loopback device gr-loopX is enslaved to this VRF.
A default route is then added in table 1000 + X, pointing to gr-loopX,
so that all traffic originating from the VRF is routed back through its
local loopback.

For the VRF 0 (i.e. gr-loop0), no Linux VRF interface is created;
instead, the default route in the main table is updated to point to
gr-loop0. This is intentional, as Grout should run inside a dedicated
network namespace.

This setup ensures correct handling of local traffic for daemons such as
bgpd that listen on 0.0.0.0 and are not bound to a specific device.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>
Signed-off-by: Robin Jarry <rjarry@redhat.com>
Reviewed-by: Christophe Fontaine <cfontain@redhat.com>