tests: add a prefix to linux interfaces
We used to have the same interface name in grout and in linux.
This was misleading as veth-pairs can't have the same name,
and the expectation is a mapping between grout and linux interfaces.

With the future control plane interfaces, there is a requirement
to have the *same* name as declared in linux for FRR. This results in
failed tests.

Prefix the tap interfaces with "x-" in the tests.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>
Reviewed-by: Robin Jarry <rjarry@redhat.com>