ecmp: add reta for routes with different weights
The basic implementation of ECMP didn't care about the weight
of each route.
A redirection table is added to the nexthop group, which is filled
with the nexthop pointers according to their respective weights.

We try to get 1 entry for each NH, but there is no guarantee if
the ratio is to high or if there are too many NH in the NHG.

The RETA has a size up to 4096 entries.

Add a new helper to get the nexthop from a nexthop group.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>
Reviewed-by: Robin Jarry <rjarry@redhat.com>