loopback: fix packet type of packet received from linux
The loopback code was checking `data[0] == 6` to decide whether a packet
was IPv6, but the IP version is stored in the upper 4 bits of the first
byte.

Use `(data[0] & 0xf0) == 0x60` to check the version field and correctly
set `mbuf->packet_type` to either `RTE_PTYPE_L3_IPV6` or `RTE_PTYPE_L3_IPV4`.

Fixes: cdfe4353139d ("loopback: rebuild ip/ip6 header")
Signed-off-by: Maxime Leroy <maxime@leroys.fr>
Reviewed-by: Christophe Fontaine <cfontain@redhat.com>
Acked-by: Robin Jarry <rjarry@redhat.com>