infra: sync iface addresses to per-VRF gr-loopX so kernel can bind
When an address is configured on an interface in Grout, automatically
add the same address to the corresponding per-VRF loopback (gr-loopX) as
a host address (/32 or /128).

This mirrors Grout-managed local addresses into the kernel namespace,
allowing Linux applications to bind()/connect() using those addresses
(per VRF) without manual IP setup.

The change wires the add/del paths so the loopback is kept in sync with
Grout’s interface addresses.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>
Reviewed-by: Christophe Fontaine <cfontain@redhat.com>
Acked-by: Robin Jarry <rjarry@redhat.com>