vrf: add default route with high metric in RT_TABLE_MAIN
The default route for vrf-0 is created inside its namespace to direct
all traffic to Grout. Normally, Grout should run in its own dedicated
namespace, where having such a default route is safe.

However, if Grout is started in the primary namespace by mistake,
this route could override the system’s default route and disrupt host
connectivity. To avoid that, configure the route with a high metric so
it is only used as a last resort.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>
Signed-off-by: Robin Jarry <rjarry@redhat.com>
Reviewed-by: Christophe Fontaine <cfontain@redhat.com>