smoke: allow external debugging of network namespaces
Since recently, the smoke tests are executed in a "grout" netns which is
reused across test runs.

Smoke tests do create additional namespaces to act as routers connected
to grout.

The issue is that creating "nested" network namespaces with iproute2 is
not possible. The file created in /run/netns has a reference which is
unusable outside of the process that has created it.

 ~# ip netns
 Error: Peer netns reference is invalid.
 Error: Peer netns reference is invalid.
 grout
 ns-a
 ns-b

 ~# ip netns exec ns-b ip link show
 setting the network namespace "ns-b" failed: Invalid argument

The only way to have namespaces which are visible/accessible from
everywhere is to create them from the base namespace (the netns of PID
1).

Fixes: 4853d4d9b129 ("smoke: make private netns work for single tests")
Fixes: 4d5fa1c6b65b ("main,smoke: run in private netns")
Signed-off-by: Robin Jarry <rjarry@redhat.com>
Reviewed-by: Maxime Leroy <maxime@leroys.fr>