frr: update version to 10.5.0
We don't need to patch FRR to build our out of tree plugin anymore. We
just need a local patch when building FRR for our local tests.

Hopefully soon, we'll have frr-headers packages that we can install from
Fedora and we'll be able to use them instead of frr for DEB and RPM
packages.

Signed-off-by: Robin Jarry <rjarry@redhat.com>
Reviewed-by: Christophe Fontaine <cfontain@redhat.com>