port: silently ignore multicast mac filters
Since commit ec4cbdc7f18f ("port: always enable allmulti and simplify
MAC filtering"), ALLMULTI is always on and adding a multicast address to
the mac filter of port causes a hard error.

This is a bit excessive. When a multicast address is added/removed
to/from a port filter, simply ignore it.

Signed-off-by: Robin Jarry <rjarry@redhat.com>
Reviewed-by: Anthony Harivel <aharivel@redhat.com>