graph: use less compact layout by default
Avoid nodes that are downstream in the processing pipeline appearing
before (i.e. on the left) of upstream nodes. For instance, ip_output
was rendered at the same level as port_rx whereas ip_input was much
further to the right. Enable GraphViz newrank algorithm by default
to produce a more intuitive left-to-right flow.

Add a compact flag to restore the previous more compact but less
intuitive layout when needed.

Signed-off-by: Robin Jarry <rjarry@redhat.com>
Reviewed-by: Christophe Fontaine <cfontain@redhat.com>