frr: compile with same c_args that pkgconfig
FRR 10.5 introduces a pkgconfig file that exposes the proper compiler
arguments for building out-of-tree plugins. Until we upgrade to that
version, we replicate the relevant c_args in our Meson build so our
plugin is built with settings consistent with FRR.

The plugin previously relied on config.h, which would normally be
needed to match FRR’s build-time feature detection. However, when adding
header installation support, FRR maintainers declined to install
config.h to avoid clashes with similarly named files from other
projects. As a result, the file is simply not available to external
plugins, and we have to build without it — even though this may expose
us to configuration mismatches.

This commit removes the config.h include.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>
Reviewed-by: Robin Jarry <rjarry@redhat.com>