frr: create a fake sr0 interface in zebra
Before FRR 10.5, seg6local routes were created on lo on the kernel side
for the default VRF. But the Linux kernel does not apply seg6local
behaviors on lo: seg6local is implemented via LWT (lightweight tunnels)
and LWTs are not executed on the loopback device (packets that hit lo go
through the local-input path only, so the LWT action is skipped).

With FRR 10.5, FRR now installs seg6local routes on a dummy interface
named "sr0" instead of lo for the default VRF. Zebra also checks for the
presence of this interface and refuses to install seg6local routes if
sr0 does not exist.

For Grout, we don't have these limitations:  However, FRR still
enforces its internal checks and won't install seg6local routes in the
default VRF unless an sr0 interface exists.

Create a fake sr0 interface in zebra so that FRR accepts seg6local
routes, even though the underlying kernel device is not needed by Grout.

A dummy sr0 device is also created in on the kernel to have an ifindex.
The ifindex of sr0 is mapped on a specific index of grout. No interface
exists on grout for this interface.

Limit:
- sr0 should normally be created via grcli so that Grout owns the
interface namespace. Creating it inside zebra instead of via grcli can
otherwise cause conflicts with Grout's own interface management.
- sr0 dummy is not destroyed at end of zebra plugin

Link: https://github.com/FRRouting/frr/commit/6b99a6c1510ff5b45771b9b79996ff38b741c50d
Signed-off-by: Maxime Leroy <maxime@leroys.fr>