github: allow changing subprojects mid-series
When a pull request contains multiple patches each patch is compiled
tested and linted using `git rebase --exec`.

If one of the patches updates the version of a subproject or introduces
a patch, the change will not be taken into account because meson has
already processed the wrap files.

Add a script that is called during the rebase operation and that will
delete the subproject folders when there was any change to its
corresponding wrap file.

Signed-off-by: Robin Jarry <rjarry@redhat.com>
Reviewed-by: Christophe Fontaine <cfontain@redhat.com>