port: disable hardware vlan filter for bifurcated drivers
When using grout with mlx5 VFs, asking to offload VLAN filter does not
return an error but it causes all unicast packets to be steered to the
kernel net device RXQs instead of the ones configured for grout.

Disabling VLAN filter offload fixes the issue. Unfortunately, there is
no way of knowing whether a given device is using a "bifurcated" driver
in a generic way. For now, rely on the driver name containing "mlx5". It
is the only driver that we know of that has this limitation so far.

Add a new features field on the iface_info_port structure. Store
a PORT_F_VLAN_FILTER flag when the port was successfully configured with
RTE_ETH_RX_OFFLOAD_VLAN_FILTER offload.

If PORT_F_VLAN_FILTER isn't available, skip adding/removing any VLAN
filter in the hardware.

Signed-off-by: Robin Jarry <robin@jarry.cc>