nexthop: introduce explicit destroy operation
When creating a nexthop, it has refcount=0. If we need to destroy it
before its refcount has been incremented we need to do the awkward
incref, decref combo.

This is confusing and we will surely forget why this is even there in
a couple of months.

Add an explicit nexthop_destroy() function and call it from within
nexthop_decref() and in the IPv6 address code.

Fixes: 991a8c0e315c ("address: fix assert error when mcast cannot be configured")
Signed-off-by: Robin Jarry <rjarry@redhat.com>
Reviewed-by: Maxime Leroy <maxime@leroys.fr>