infra: add helper gr_mbuf_prepend
rte_pktmbuf_prepend returns NULL when there is not enough headroom.
The first possible option would be to go with multiseg.
As we may not want to go with it today, try as much as possible to
return a valid mbuf, even at the cost of performances.

In the case where a prepend fails (with a very large SRv6 header),
move the data forward to make space.

Forbid usage of rte_pktmbuf_prepend in favor of gr_mbuf_prepend.

Suggested-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: Christophe Fontaine <cfontain@redhat.com>
Reviewed-by: Robin Jarry <rjarry@redhat.com>