vlan: add set_up_down callback
Implement the set_up_down callback for VLAN interfaces. When bringing
the interface up, only set the running state and push a status up event
if the parent interface is also running. When bringing it down, clear
both the up flag and running state, then push a status down event.

Signed-off-by: Robin Jarry <rjarry@redhat.com>
Reviewed-by: Christophe Fontaine <cfontain@redhat.com>