vlan: fix state inheritance on parent change
When a VLAN interface is moved to a different parent during
reconfiguration, clear the running state before removing it from the
old parent. Do not blindly copy the new parent's state. Instead, only
set running state and push a status up event if the interface has the
up flag set, the parent is running, and we are not also modifying the
up flag in this reconfiguration.

Signed-off-by: Robin Jarry <rjarry@redhat.com>
Reviewed-by: Christophe Fontaine <cfontain@redhat.com>