ip_error: avoid accessing ip header after prepend
If there is not enough headroom in the mbuf, gr_mbuf_prepend() may
memmove() the data to recover some headroom. We must not access the IP
header pointer after that.

Since we only need the source address and IP length, store them in
variables before prepending.

To avoid confusion, reuse the src variable to store the source address
and add a new dst variable to store the destination gateway used to
determine which local address should be used as source.

Signed-off-by: Robin Jarry <rjarry@redhat.com>