bond: remove port from members on destroy
When deleting a port interface which is part of a bond, a dangling
pointer is left in the bond members array. This causes uses-after-free
issues accessing bond members:

AddressSanitizer: heap-use-after-free on address 0x7b85fbe15100
READ of size 2 at 0x7b85fbe15100 thread T0
    #0 bond_to_api ../modules/infra/control/bond.c:399
    #1 iface_to_api ../modules/infra/api/iface.c:24
    #2 iface_list ../modules/infra/api/iface.c:100

Make sure to remove the member when a port is deleted. Update the
primary index accordingly.

Signed-off-by: Robin Jarry <rjarry@redhat.com>